refactor : 수익 결정하는 메서드 클래스 분리
